hashtag i hate size_t

+++ vojuck.victoria/P4/findMaxSum.hpp
@@ -1,6 +1,6 @@
 #ifndef FINDMAXSUM_HPP
 #define FINDMAXSYM_HPP
 
-int findMaxSum(int* array, int rows, int cols);
+int findMaxSum(int* array, size_t rows, size_t cols);
 
 #endif
