+++ vojuck.victoria/P4/isUpperTriangular.hpp
@@ -1,6 +1,6 @@
 #ifndef ISUPPERTRIANGULAR_HPP
 #define ISUPPERTRIANGULAR_HPP
 
-bool isUpperTriangular(int* matrix, int rows, int cols);
+bool isUpperTriangular(int* matrix, size_t rows, size_t cols);
 
 #endif
